FAT-1703 Created static method DogsDataSource.newInstance. GSON serializing logic was replaced in DogsDataSource class
