+++ app/src/main/java/com/andersen/dogsapp/dogs/DogsDataSource.java
@@ -1,6 +1,13 @@
 package com.andersen.dogsapp.dogs;
+import android.content.Context;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
+
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -27,7 +34,34 @@ public class DogsDataSource {
         return dogsDataSource;
     }
 
+    public static DogsDataSource getInstance (Context context){
+        GsonBuilder builder = new GsonBuilder();
+        final Gson GSON = builder.create();
+        String json;
+        json = getAssetsJSON("dogs.json", context);
+
+        if(dogsDataSource == null){
+            dogsDataSource = new DogsDataSource(GSON.fromJson(json, DogsDataSource.class));
+        }
+        return dogsDataSource;
+    }
+
     public List<Dog> getDogs(){
         return dogs;
+    }
+
+    private static String getAssetsJSON(String fileName, Context context){
+        String json = null;
+        try{
+            InputStream inputStream = context.getAssets().open(fileName);
+            int size = inputStream.available();
+            byte[] buffer = new byte[size];
+            inputStream.read(buffer);
+            inputStream.close();
+            json = new String (buffer, "UTF-8");
+        } catch (IOException e ){
+            e.printStackTrace();
+        }
+        return json;
     }
 }
